Fix request changes
Change endpoints from /Organizations to /organizations
Minor tweaks
